requirements.txt
# Web framework
Flask==2.3.3
Werkzeug==2.3.7

# Supabase (stack compatível)
supabase==2.4.0
postgrest==0.16.0

# HTTP client (FIXADO para evitar conflitos)
httpx>=0.24,<0.26

# Environment variables
python-dotenv==1.0.1

# Auth / session
Flask-Login==0.6.3

